first refactorings
